feat: implement user table filtering components with role and product selection support
